Fixed #42 prevent files deletion when control is disabled

+++ projects/iplab/ngx-file-upload/src/lib/helpers/control.class.ts
@@ -105,9 +105,11 @@ export class FileUploadControl {
     }
 
     public removeFile(file: File): this {
-        this.files.delete(file);
-        this.validate();
-        this.valueChanges.next(Array.from(this.files.values()));
+        if (!this.disabled) {
+            this.files.delete(file);
+            this.validate();
+            this.valueChanges.next(Array.from(this.files.values()));
+        }
         return this;
     }
